requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudscraper==1.2.71
et-xmlfile==1.1.0
Flask==3.0.3
h11==0.14.0
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.4
lml==0.1.0
lxml==5.2.2
MarkupSafe==2.1.5
openpyxl==3.1.4
outcome==1.3.0.post0
pyexcel-ezodf==0.3.4
pyexcel-io==0.6.6
pyexcel-ods3==0.6.1
pyexcel-xlsx==0.6.0
pyparsing==3.1.2
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
wsproto==1.2.0
